pyproject.toml
#
# UMCP runnable + publication-grade repo configuration.
# - build system (setuptools)
# - project metadata
# - runtime deps (minimal)
# - dev/test deps
# - pytest config
#
# IMPORTANT:
# - keywords belongs at [project].keywords (list of strings)
# - [project.urls] values MUST be strings (valid URLs), not lists

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "umcp"
version = "1.3.2"
description = "Universal Measurement Contract Protocol (UMCP): contract-first validation, receipts, and runnable casepacks with GCD and RCFT frameworks. Core Axiom: What Returns Through Collapse Is Real (no_return_no_credit). Features intelligent caching, progressive optimization, and extensible plugin architecture."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

authors = [
  { name = "Clement Paulus" }
]

keywords = ["UMCP", "GCD", "RCFT", "contract-first", "validation", "reproducibility", "casepacks", "fractal", "recursive-field"]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: System :: Monitoring"
]

# Runtime dependencies:
# Keep runtime small; tests/dev tooling goes in optional-dependencies.
dependencies = [
  "pyyaml>=6.0.1",
  "jsonschema>=4.23.0",
  "numpy>=1.24.0",
  "scipy>=1.10.0"
]

# Production-grade optional dependencies
[project.optional-dependencies]
production = ["psutil>=5.9.0"]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-xdist>=3.8.0",
  "ruff>=0.5.0",
  "mypy>=1.11.0",
  "pre-commit>=3.8.0",
  "psutil>=5.9.0"
]
test = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-xdist>=3.8.0"
]
# Extensions: visualization and API
viz = [
  "streamlit>=1.30.0",
  "pandas>=2.0.0",
  "plotly>=5.18.0"
]
api = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0"
]
# All extensions
extensions = [
  "streamlit>=1.30.0",
  "pandas>=2.0.0",
  "plotly>=5.18.0",
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0"
]

# These MUST be strings; do not put lists here.
[project.urls]
Homepage = "https://github.com/calebpruett927/UMCP-Metadata-Runnable-Code"
Repository = "https://github.com/calebpruett927/UMCP-Metadata-Runnable-Code"
Issues = "https://github.com/calebpruett927/UMCP-Metadata-Runnable-Code/issues"

[project.scripts]
umcp = "umcp.cli:main"
umcp-visualize = "visualize_umcp:main"
umcp-api = "api_umcp:run_server"
umcp-ext = "umcp_extensions:main"
umcp-format = "umcp_autoformat:main"

# Extension discovery and plugin system
[project.entry-points."umcp.extensions"]
visualization = "visualize_umcp:UMCPVisualization"
api = "api_umcp:UMCPAuditAPI"
ledger = "umcp.cli:LedgerExtension"

[project.entry-points."umcp.contracts"]
uma = "contracts.UMA.INTSTACK.v1:contract"
gcd = "contracts.GCD.INTSTACK.v1:contract"
rcft = "contracts.RCFT.INTSTACK.v1:contract"

[project.entry-points."umcp.closures"]
gcd_closures = "closures.gcd:register"
rcft_closures = "closures.rcft:register"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["umcp*"]

[tool.setuptools.package-data]
umcp = [
  "py.typed",
  "../contracts/*.yaml",
  "../closures/**/*.py",
  "../schemas/*.json",
  "../canon/*.yaml"
]

# Auto-discovery of extensions
[tool.setuptools.dynamic]
# Extensions are auto-discovered from root directory

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
  "default",
  "ignore::DeprecationWarning"
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

[tool.coverage.run]
branch = true
source = ["src/umcp"]
omit = [
  "src/umcp/cli.py",  # CLI is subprocess-heavy, difficult to test comprehensively
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "I",    # isort
  "B",    # flake8-bugbear
  "UP",   # pyupgrade
  "SIM",  # flake8-simplify
  "C4",   # flake8-comprehensions
  "RUF",  # Ruff-specific rules
]
ignore = [
  "E501",    # Line too long (handled by formatter)
  "RUF001",  # Ambiguous unicode in strings (intentional for Greek letters)
  "RUF002",  # Ambiguous unicode in docstrings (intentional for Greek letters)
  "RUF003",  # Ambiguous unicode in comments (intentional for Greek letters)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501", "SIM105"]  # Allow longer lines and try-except-pass in tests
"scripts/**" = ["E501"]  # Allow longer lines in scripts
"src/umcp/__init__.py" = ["E402"]  # Module imports after version check

# Type checking configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
disallow_any_generics = true
disallow_subclassing_any = true
warn_return_any = true
warn_unreachable = true
check_untyped_defs = true

# Module-specific overrides for third-party untyped libraries
[[tool.mypy.overrides]]
module = [
  "yaml.*",
  "streamlit.*",
  "plotly.*",
  "uvicorn.*",
]
ignore_missing_imports = true